Extract common shadow style
