+++ xcode/Subconscious/Shared/Components/Detail/MemoEditorDetail.swift
@@ -230,12 +230,7 @@ struct MemoEditorDetailView: View {
                         )
                         .background(background)
                         .cornerRadius(DeckTheme.cornerRadius, corners: .allCorners)
-                        .shadow(
-                            color: DeckTheme.cardShadow.opacity(0.08),
-                            radius: 1.5,
-                            x: 0,
-                            y: 1.5
-                        )
+                        .shadow(style: .transclude)
                         .padding(.bottom, AppTheme.unit4)
                         .padding(.top, AppTheme.unit2)
